Move number-to-string function to utility.*
